Isolate fixing routines and pull out of `segments`.
