+++ setup.cfg
@@ -164,3 +164,34 @@ source_modules =
     sqlfluff.api
 forbidden_modules =
     sqlfluff.cli
+
+[importlinter:contract:helper-interdependence]
+name = Helper methods must be internally indepentent
+type = independence
+modules =
+    sqlfluff.core.helpers.string
+    sqlfluff.core.helpers.slice
+
+[importlinter:contract:core-layers]
+name = Dependency layers within core
+# NOTE: Several modules within core currently have somewhat more
+# convoluted dependency loops, especially when it comes to type checking.
+# Those are currently excluded from this work, but might be picked up in
+# future work to help with better isolation.
+type = layers
+layers=
+    # `linter` references many things, including rules.
+    sqlfluff.core.linter
+    # `rules` should be independent from linter, but can reference the others.
+    sqlfluff.core.rules
+    # `parser` should be independent of `rules` and `linter`.
+    sqlfluff.core.parser
+    # `errors` should be a utility library, which can be referenced by the others.
+    sqlfluff.core.errors
+    # `helpers` should be independent and not reference any of the above.
+    sqlfluff.core.helpers
+ignore_imports =
+    # `errors` references `rules` and `parser`, but only for type checking.
+    # Splitting it up seems overkill for now, so an allowable exception.
+    sqlfluff.core.errors -> sqlfluff.core.rules
+    sqlfluff.core.errors -> sqlfluff.core.parser
